refactor optimization reporter system (#25231)
