+++ modules/optimization/include/optimizationreporters/OptimizationReporter.h
@@ -9,6 +9,7 @@
 
 #pragma once
 
+#include "OptimizationDataHelper.h"
 #include "OptimizationReporterBase.h"
 
 /**
@@ -16,10 +17,27 @@
  */
 class OptimizationReporter : public OptimizationReporterBase
 {
+private:
+  /// Data helper for generating objective value.
+  OptimizationDataHelper _opt_data;
+
 public:
   static InputParameters validParams();
   OptimizationReporter(const InputParameters & parameters);
 
+  void execute() override;
+
+  virtual Real computeObjective() override;
+  virtual void setMisfitToSimulatedValues() override;
+
+  /// measurement values
+  std::vector<Real> & _measurement_values;
+  /// simulated values at measurement xyzt
+  std::vector<Real> & _simulation_values;
+  /// difference between simulation and measurement values at measurement xyzt
+  std::vector<Real> & _misfit_values;
+
 private:
-  std::vector<Real> fillParamsVector(std::string type, Real default_value) const;
+  virtual void setICsandBounds() override;
+  virtual void setSimulationValuesForTesting(std::vector<Real> & data) override;
 };
