+++ modules/optimization/include/reporters/OptimizationData.h
@@ -8,7 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #pragma once
-
+#include "OptimizationDataHelper.h"
 #include "GeneralReporter.h"
 /**
  * Reporter to hold measurement and simulation data for optimization problems
@@ -25,32 +25,6 @@ public:
   virtual void finalize() override {}
 
 protected:
-  ///@{
-  /// x,y,z,t and value measurement data
-  std::vector<Real> & _measurement_xcoord;
-  std::vector<Real> & _measurement_ycoord;
-  std::vector<Real> & _measurement_zcoord;
-  std::vector<Real> & _measurement_time;
-  std::vector<Real> & _measurement_values;
-  ///@}
-  /// simulated values at measurment xyzt
-  std::vector<Real> & _simulation_values;
-  /// difference between simulation and measurment values at measurment xyzt
-  std::vector<Real> & _misfit_values;
-
 private:
-  /// parse measurement data from csv file
-  void readMeasurementsFromFile();
-  /// parse measurement data from input file
-  void readMeasurementsFromInput();
-  /// private method for testing optimizationData with test src
-  void setSimulationValuesForTesting(std::vector<Real> & data);
-  /// variable
-  std::vector<MooseVariableFieldBase *> _var_vec;
-  /// Weight names to reporter values
-  std::vector<std::vector<Real> *> _variable_weights;
-  /// Weight names to reporter values map created from input file
-  std::map<std::string, std::vector<Real> *> _weight_names_weights_map;
-  /// helper to check data sizes
-  void errorCheckDataSize();
+  OptimizationDataHelper _opt_data;
 };
