+++ modules/optimization/src/optimizationreporters/ParameterMeshOptimization.C
@@ -7,6 +7,7 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
+#include "OptimizationReporterBase.h"
 #include "ParameterMeshOptimization.h"
 
 #include "AddVariableAction.h"
@@ -19,6 +20,8 @@ InputParameters
 ParameterMeshOptimization::validParams()
 {
   InputParameters params = OptimizationReporterBase::validParams();
+  params += OptimizationDataHelper::validParams();
+
   params.addClassDescription(
       "Computes objective function, gradient and contains reporters for communicating between "
       "optimizeSolve and subapps using mesh-based parameter definition.");
@@ -73,7 +76,11 @@ ParameterMeshOptimization::validParams()
 }
 
 ParameterMeshOptimization::ParameterMeshOptimization(const InputParameters & parameters)
-  : OptimizationReporterBase(parameters)
+  : OptimizationReporterBase(parameters),
+    _opt_data(parameters),
+    _measurement_values(_opt_data.getMeasurementValues()),
+    _simulation_values(_opt_data.getSimValues()),
+    _misfit_values(_opt_data.getMisfitValues())
 {
 
   _nvalues.resize(_nparams, 0);
@@ -255,4 +262,42 @@ ParameterMeshOptimization::parseData(const std::vector<unsigned int> & exodus_ti
     parsedData.resize(parsedData.size() + numberOfControllableParameters, constantDataFromInput);
 
   return parsedData;
+}
+
+void
+ParameterMeshOptimization::execute()
+{
+  _opt_data.computeMisfit();
+}
+
+Real
+ParameterMeshOptimization::computeObjective()
+{
+  // This will only be executed if measurement_values are available on the main app
+  for (const auto i : index_range(_measurement_values))
+    _misfit_values[i] = _simulation_values[i] - _measurement_values[i];
+
+  Real val = 0.0;
+  for (auto & misfit : _misfit_values)
+    val += misfit * misfit;
+
+  if (_tikhonov_coeff > 0.0)
+  {
+    Real param_norm_sqr = 0;
+    for (const auto & data : _parameters)
+      for (const auto & val : *data)
+        param_norm_sqr += val * val;
+
+    val += _tikhonov_coeff * param_norm_sqr;
+  }
+
+  return val * 0.5;
+}
+
+// function only used for test objects
+void
+ParameterMeshOptimization::setSimulationValuesForTesting(std::vector<Real> & data)
+{
+  _simulation_values.clear();
+  _simulation_values = data;
 }
